Removed the use of Z, used Y to shift both lines
Using Z was only hurting us. Since we were trying to shift the Z memory as well as the Y memory we were essentially trying to pull from the prog and data mem at the same time and it was only getting messy. This should hopefully start at the beginning of the data mem with Y pointing to 0x0100 ("W" in the first line) and then move 30 spaces to the bottom (may need to be 32 since the strings are 16 in length). Once there it will save that last char. It will then enter the shift loop. Within the shift loop there is a predecrement to grab one char to the left (one up in vertical memory representation). Y is then incremented by 1 to shift to the new location of the grabbed char and stores it there. At the end the Y is decremented once because it will be decremented again at the start of the loop so it will now look to grab the next character to the left/up to be moved right/down. It will conclude once the counter reaches 0 and then store the last char into the top memory location/first char of line 1.

If more garbage is produced from this it could just mean we need one less padding char on our strings because I think we only have 31 memory locations in the data mem to work with so the strings may both need to be 15 in length rather than 16 for this to not try and pull memory from nowhere.